+++ src/codegen/irgen/util.h
@@ -28,9 +28,9 @@ namespace pyston {
 llvm::Constant* getStringConstantPtr(const std::string& str);
 llvm::Constant* getStringConstantPtr(const char* str);
 llvm::Constant* embedConstantPtr(const void* addr, llvm::Type*);
-llvm::Constant* getConstantInt(int val);
+llvm::Constant* getConstantInt(int64_t val);
 llvm::Constant* getConstantDouble(double val);
-llvm::Constant* getConstantInt(int val, llvm::Type*);
+llvm::Constant* getConstantInt(int64_t val, llvm::Type*);
 
 void dumpPrettyIR(llvm::Function* f);
 }
